CV2-4231 Fixes and additions to shared feed confirmation modals (#1906)
* Added modal for creating feed with no invites and edited all other confirmation modals

Updated copy for creating and editing feeds
Added an alert to inform the user that updates are not in real time
Added workspace name to confirmation message. If creating, it's the workspace of the user.  If editing it's the feed's workspace.
Changed save function to always prompt the user to confirm in the modal before saving

CV2-4231 - Fixes and additions to shared feed confirmation modals

* Updated copy to match spec when updating feed