Refinement, PR fix requests
Added the already existing Cancel label.
Improved the alignment of footer objects and wrapped buttons for smaller layouts.
Added props for Cancel button (cancelProps) so it can be edited to match the style of the other buttons
Added gap to footer buttons. Removed unnecessary uses of cx. Replaced right and left changes with translate3d
Long titles are now truncated with an ending ellipses.