CV2-4477 new Slideout component (#1923)
* Added a new slideout component

Slideout component requires a title (object or string) and an onClose function.
Footer and all its contents are optional.
Contents can be Close, MainActionButton, SecondaryActionButton, and OptionalNode.
Content can contain anything and is scrollable.
Slideout works in rtl and ltr layouts.

DEV-4477

* Refinement, PR fix requests

Added the already existing Cancel label.
Improved the alignment of footer objects and wrapped buttons for smaller layouts.
Added props for Cancel button (cancelProps) so it can be edited to match the style of the other buttons
Added gap to footer buttons. Removed unnecessary uses of cx. Replaced right and left changes with translate3d
Long titles are now truncated with an ending ellipses.

* Fixes to css for small amounts of content