Add MetadataText from check-ui to check-web (#1928)
* move MetadataText component from check-ui to check-web

* CV2-4477 new Slideout component (#1923)

* Added a new slideout component

Slideout component requires a title (object or string) and an onClose function.
Footer and all its contents are optional.
Contents can be Close, MainActionButton, SecondaryActionButton, and OptionalNode.
Content can contain anything and is scrollable.
Slideout works in rtl and ltr layouts.

DEV-4477

* Refinement, PR fix requests

Added the already existing Cancel label.
Improved the alignment of footer objects and wrapped buttons for smaller layouts.
Added props for Cancel button (cancelProps) so it can be edited to match the style of the other buttons
Added gap to footer buttons. Removed unnecessary uses of cx. Replaced right and left changes with translate3d
Long titles are now truncated with an ending ellipses.

* Fixes to css for small amounts of content

* [smoke tests] Run smoke tests for this branch

* [smoke tests] Run smoke tests for this branch

* Revert " CV2-4477 new Slideout component (#1923)"

This reverts commit 098e4595ef3bf74b82d1911013883a370b5c70ca.

* [smoke tests] Run smoke tests for this branch

* [smoke tests] Run smoke tests for this branch

* [smoke tests] Run smoke tests for this branch

* [smoke tests] Run smoke tests for this branch

* update test to remove clear-button no longer presetn [smoke tests]

* target different clear button for testing [smoke tests]

* put back missing test class selector [smoke tests]

* look for correct dom element with edited annotation title [smoke tests]

* replace some material buttons and styled components while fixing tests [smoke tests]

* cleanup layout on helpers error messaging to use css

* use some global translated strings inside of the task component

* add json file translation updates

---------

Co-authored-by: Sarah Kehoe <161879347+sarahkeh@users.noreply.github.com>