Added creation forms for Explainers and Fact Checks (#1960)
* Added creation forms for Explainers and Fact Checks

ExplainerForm handles the mutations for creating or updating an explainer article. The explainer article requires title, summary, and language.  Tags and url are optional.
ClaimAndFacCheckForm handles the  mutations for creating or upating a fact check.  A claim description is first required before removing the overlay for the Fact Check section.
Form components were mostly taken from the MediaFactCheck component. An update was made to the LnaguagePicker to allow for the field to be required for the explainer.

cv2-4498 and cv2-4447

* Connect forms with links on the Articles page, made necessary updates to styling and passed down team info needed for saving any new Form

* Request pr changes for input forms

Opening another slideout from the New Article menu will close the current slideout.
Menu will close when a selection is made.
Flash message now uses the useContext hook
Changed error handling failed saves, will use the API error and fallback on a generic message
Specified the possible strings for articleType and mode values
Removed the reliance on values for TranslatedMessages which was preventing the messages from truly being translated
Removed error value that was left over from testing

* removed console statement

* Added team tags for the forms, made additional fixes to simplify value assignments.  Added tag_texts to Projects query since the New Article button lives within that object

* saving the wrong info