Make sure that weekdays displayed in newsletter scheduler are the same regardless the current user timezone.
Previously, the implementation was using `new Date(x, y, z)`, which takes into account the current user timezone. So, for a user browsing using timezone Asia/Calcutta, for example, the weekdays displayed in the newsletter scheduler started by Monday. For another timezone, for example America/New_York, the weekdays started by Sunday. The fix is to pass the UTC timezone to `new Date()`. This PR also includes unit tests that reproduce this issue.

Fixes: CV2-4845.