Tests/CV2-4883: Fix integration tests (#2022)
Some of the tests that failed were related to changes in the UI. For example, the "Accept/Reject Similarity" button used to be available for each media item, but now it only appears when hovering over the item. Consequently, the tests failed due to the "element not interactable" error. Additionally, some tests failed because the action was performed, but the auto-refresh did not occur to reflect the action on the page, To address this, I added some @driver.navigate.refresh on the tests.

Reference: CV2-4883