Should be able to create claims without fact-checks. (#2086)
Disclaimer: I don't like the solution that I implemented here, tbh. But it was the simplest thing I could think of. Reminding the idea of the ticket (CV2-5134): do the least work possible that doesn't move us backward in order to (hopefully temporarily) support a single partner that still uses claims without fact-checks. So, I did here something that doesn't require any reversion in the backend, and that's actually what we did previously, after the release, to migrate claims without fact-checks to the new architecture. In a nutshell, create a fact-check with placeholder values, that are not indexed. Sub-tasks:

- Pre-select default language in articles forms
- Fix "unknown language" bug
- These fields should not be required anymore: fact-check title, fact-check summary and language
- Should be able to create claim without fact-check
- If fact-check fields are not ready, the "Unpublished Report" button should be disabled and display a tooltip
- In the articles cards, display claim title and context if there is no fact-check title and summary
- Articles list
- Applied article card in item page
- "Recent articles" list in item page

Reference: CV2-5134.