make the number column a fixed width to 7 chars
It was hard to "scan" the list since the text could end in different places.

By having the number area be a fixed width that can accept up to 7 digits,
which is the max number we will probably see in the system, the text column
will always "end" at the same place, so it will be easier to visually run
the list.

https://github.com/meedan/check-web/pull/2114#pullrequestreview-2315777643