make sure missing elements don't affect the DOM
For optional elements, its typically better to not have them influence
the DOM if they are not present, since it can alter the layout.

Before when we left out eft out the contextText we could still see the
<div.. that outputs it is still written to the screen. You can’t see
anything visually, but what it does is make the widget artificially
taller than it should be since it still takes up space.

When we don't add the div the padding of the widget is “correct”