5449 – Update NumberWidget and ListWidget to be dynamic (#2158)
When I added the Number and List Widget I just added static different variations. Meaning the same component was added a few times.

Since we will be adding more components to the Sandbox, it will get longer. So we should be able to change components with variations by changing options. This means: switches and selects to choose what optional items to show and hide.

Sidenote: This also made a small change to the sandbox navigation.

References: 5449, 4933, 4934
PR: 2158