Add time to request status tickers (#2164)
## Description

We need to display the time along with the date in the request status tickers. However, since the TimeBefore component is reused in multiple places across the codebase, adding the includeTime prop allows us to extend its functionality without affecting other areas where we only want to show the date. This change ensures flexibility by including the time where necessary while maintaining the existing behavior in components that don't require it.

-  [X] display the time along with the date in request status tickers.
-  [X] add a new includeTime prop to control whether or not the time is shown.
-  [X] Add unit tests
-  [X] remove /* eslint-disable react/sort-prop-types */ from the component

Reference: CV2-4050