Correctly format the header first and last date
Since now we send a hyphen as a value for the date, the dates in the
table header were being displayed as three hyphens '---'.
So we change this to not display anything if we don't have valid dates,
if we only have only valeu in the dates array and it's an hyphen.

relates to:
- https://github.com/meedan/check-api/pull/2143
- CV2-5607