"All Articles" list. (#2315)
This PR adds a new article list called "All Articles." As the name suggests, it includes all articles (fact-checks and explainers) and serves as the default option when "Articles" is clicked in the sidebar.

As part of this work, this PR also significantly refactors the article list components. The logic for update mutations and sorting options, which was previously duplicated across multiple components, has been moved to the abstract "Articles.js" component.

Reference: CV2-5007.