Refactoring/CV2-5203: Return all items that match a search (#2332)
Description
Previously, when a fact-check was linked to another item, it was excluded from search results. We modified the query parameter to return all media types to ensure that items appear in search results along with their associated fact-checks. For the fact-check tab, it was only returning items of blank type:
AutoCompleteMediaItem.js#L105.

Additionally, added a tooltip to inform users that the fact-check is not applicable and disabled the item to prevent selection.

Reference: CV2-5203