fix and run [smoke tests] (#2349)
Fix test:
This test was broken due to recent changes introduced in CV2-6325, where verification statuses are now sorted alphabetically. The test was still asserting based on the previous (unsorted) order, which caused it to fail. This fix updates the test to reflect the new, correct alphabetical order of statuses.
References: CV2-6325