move content and title to FeedDataPointsSection
I think now that:
- we have the listType prop,
- we know we are not adding any new types of saved searches beyond article and media,
- and because the content and title themselves are not dynamic

We can use listType to determine which content and title to display, and
move the content and title inside the FeedDataPointsSection component.

I think this makes SaveFeed a little bit easier to read.
I also think it's easier to find the content and title inside the component
instead of having to find where the prop is being passed from.

But I think this only makes sense because we now know we don't need to
be as flexible.