5217, 6373 – Shared Feed: refactoring, copy, and design updates (#2364)
This works updates the creation of a shared feed, so we are able to select and save a media list and/or an article list. This only updates the creation of the feed, the display of both lists in the Feed will be a separate ticket. This should not break the current Feed view.

1. On design changes
We are following the new design in jira. So that article and media FeedDataPointsSection both have their own FeedContent component, where only article lists or media lists will be displayed, we had to:
    - update how we structure the components,
    - update some components to accept a listType prop,
    - update copy in the components

Structure: Before
We had the two FeedDataPointsSection, inside FeedDataPoints, which was inside SaveFeed. While FeedContent was inside SaveFeed.

Structure: After
I moved FeedContent into FeedDataPointsSection, which was moved directly into SaveFeed. I had to add a new FeedDataPointsSectionHeader with the title and alert. I did this in order to try to avoid too much redirection.

Because we don't need to be as flexible:
- we know we are not adding any new types of saved searches beyond article and media,
- and because the content and title themselves are not dynamic
We can use listType to determine which content and title to display, and move the content and title inside the FeedDataPointsSection component.

2. Article saved search
Currently, we only save one reference to a media_saved_search in shared feeds. We are updating the backend, concurrently to this work (meedan/check-api#2310), to be able to save one reference to an article_saved_search as well.

We updated so we are able to:
    - select two lists if needed (one for media and one for articles)
    - query, create, update, delete both lists

References: CV2-5217, CV2-6373