Clarify alert for explainers when media already has a fact-check attached (#2371)
Improve the alert message to clarify to the users that the explainer will not be sent because a fact-check already exists
Reference: CV2-6385

How to test?
Go to a media item that already has a fact-check.
Attempt to add an explainer.
Confirm that the updated alert message appears as described above.