setIsEditing to false when saving feed team (#2377)
When saving as an invited team, the following notice would pop-up:
'Do you want to leave without saving? You have unsaved changes to your shared feed. …'

This happened because even though we called handleSaveFeedTeam successfully,
and the feed was saved, the `isEditing` state was not set to false.
Which caused NavigateAwayDialog to pop up.

I think as part of the save operation, we should also set `isEditing` to false,
if someone is saving, they are no longer editing.

References: CV2-6419