Fix Inconsistent request count display in child item requests tab (#2379)
## Description

When an item is confirmed as a child of another item, the requests count displayed in the "Requests" tab was inconsistent with the number of requests actually assigned to that child item.

The Fix: Updated the logic to ensure that the "Requests" tab on a child item displays only the count of requests directly assigned to that child item (requests_count).
The demand count (demand) continues to be used where it applies to the parent or main item view.

requests_count: Tipline requests count assigned to an item
demand: Tipline requests count assigned to an item and it's related items

References: CV2-5421