Update bulk delete tag message (#2400)
We updated the text because there was some confusion that it should not be possible to remove a tag that is not tagged in the item.

Since we are bulk deleting: item 1 might have a, b tags and item 2 might have b, c. Lets we bulk remove b, c. c is not present in item 1. It would start getting too complex if we wanted to keep track of what was removed from each item.

So we are hoping to clarify that items without those tags will remain unaffected.

References: CV2-2955