Remove link_preview option from header type dropdown when not in availableHeaderTypes (#2381)
## Description
Filter `'link_preview'` to only be displayed when available in availableHeaderTypes.This change removes 'link_preview' from the Newsletter page, while still allowing it to appear in the New Tipline Resource section

Reference: CV2-5725 
## How to test?

Navigate to the Newsletter Settings page.
Open the Content Header dropdown menu.
Confirm that the "Link Preview" option is not displayed in the dropdown.

Navigate to New Tipline Resource
Open the Content Header dropdown menu.
Confirm that the "Link Preview" option is displayed in the dropdown.